adjust clipsize slightly to not crowd the crosshair
Former-commit-id: 01109cdff35bda944e00fb5171baf8d81d25a3e2

+++ src/game/game.h
@@ -1279,9 +1279,9 @@ namespace hud
 {
 	extern int hudwidth, hudsize;
 	extern float radarblipblend, radarnameblend, inventoryblend, inventoryskew;
-	extern void drawquad(int x, int y, int w, int h, float tx1 = 0, float ty1 = 0, float tx2 = 1, float ty2 = 1);
-	extern void drawtex(int x, int y, int w, int h, float tx = 0, float ty = 0, float tw = 1, float th = 1);
-	extern void drawsized(int x, int y, int s);
+	extern void drawquad(float x, float y, float w, float h, float tx1 = 0, float ty1 = 0, float tx2 = 1, float ty2 = 1);
+	extern void drawtex(float x, float y, float w, float h, float tx = 0, float ty = 0, float tw = 1, float th = 1);
+	extern void drawsized(float x, float y, float s);
 	extern void drawblip(int w, int h, int s, float blend, int idx, vec &dir, float r = 1.f, float g = 1.f, float b = 1.f, const char *font = "radar", float fade = -1.f, const char *text = NULL, ...);
 	extern int drawitem(const char *tex, int x, int y, float size, float fade, float skew, const char *font = NULL, float blend = 1.f, const char *text = NULL, ...);
 	extern const char *flagtex(int team = TEAM_NEUTRAL);
